requirements.txt
flask>=3.0.0
pyyaml>=6.0
psycopg2-binary>=2.9.9
